autotest_selenium_modal/requirements.txt
allure-pytest>=2.8.6
cloudinary>=1.19
pillow>=6.2.1
PyHamcrest>=1.9.0
pytest>=3.6.1
requests>=2.18.4
selenium>=3.14.0
visual-logging>=1.0
webdriver-manager>=2.3